# core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A high-performance proxy rotation server with health monitoring and intelligent routing\nProvides comprehensive API for managing proxy servers, monitoring their health,\nand configuring rotation strategies.",
        "title": "Rota Proxy API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/alpkeskin/rota"
        },
        "license": {
            "name": "LICENSE",
            "url": "https://github.com/alpkeskin/rota/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8001",
    "basePath": "/api/v1",
    "paths": {
...
        "/settings": {
            "get": {
                "description": "Get current system configuration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get settings",
                "responses": {
                    "200": {
                        "description": "Current settings",
                        "schema": {
                            "$ref": "#/definitions/models.Settings"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update system configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update settings",
                "parameters": [
                    {
                        "description": "Updated settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Settings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update confirmation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/settings/reset": {
            "post": {
                "description": "Reset configuration to default values",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Reset settings",
                "responses": {
                    "200": {
                        "description": "Reset confirmation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AuthenticationSettings": {
            "type": "object",
            "properties": {
                "enabled": {
                    "description": "Enable authentication for proxy requests",
                    "type": "boolean"
                },
                "password": {
                    "description": "Password for proxy authentication (write-only)",
                    "type": "string"
                },
                "username": {
                    "description": "Username for proxy authentication",
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "models.GeoIPSettings": {
            "type": "object",
            "properties": {
                "auto_update": {
                    "description": "Enable automatic DB update on interval",
                    "type": "boolean"
                },
                "last_updated_at": {
                    "description": "Timestamp of last successful update",
                    "type": "string"
                },
                "maxmind_db_path": {
                    "description": "Path to local .mmdb file",
                    "type": "string"
                },
                "maxmind_license_key": {
                    "description": "MaxMind license key for auto-download",
                    "type": "string"
                },
                "maxmind_url": {
                    "description": "Custom download URL (optional)",
                    "type": "string"
                },
                "provider": {
                    "description": "\"ip-api\" or \"maxmind\"",
                    "type": "string"
                },
                "update_interval_hours": {
                    "description": "How often to update in hours (default 168 = 7 days)",
                    "type": "integer"
                }
            }
        },
        "models.GlobalHealthCheckSettings": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "interval_minutes": {
                    "type": "integer"
                }
            }
        },
        "models.HealthCheckSettings": {
            "type": "object",
            "properties": {
                "headers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "idle_ttl_minutes": {
                    "description": "IdleTTLMinutes is the same TTL filter for idle-orphan sweeps.",
                    "type": "integer"
                },
                "orphan_ttl_minutes": {
                    "description": "OrphanTTLMinutes excludes orphan proxies checked more recently than now()-TTL from periodic orphan sweeps (0 disables the filter).",
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "strict_tls": {
                    "description": "Enable real TLS certificate validation during health checks",
                    "type": "boolean"
                },
                "timeout": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "workers": {
                    "type": "integer"
                }
            }
        },
...
        "models.LogRetentionSettings": {
            "type": "object",
            "properties": {
                "cleanup_interval_hours": {
                    "description": "How often to run cleanup (1, 6, 12, 24)",
                    "type": "integer"
                },
                "compression_after_days": {
                    "description": "Compress logs older than X days (1, 3, 7, 14)",
                    "type": "integer"
                },
                "enabled": {
                    "description": "Enable automatic log cleanup",
                    "type": "boolean"
                },
                "retention_days": {
                    "description": "Days to keep logs (7, 15, 30, 60, 90)",
                    "type": "integer"
                }
            }
        },
...
        "models.ProxyCleanupSettings": {
            "type": "object",
            "properties": {
                "cleanup_interval_hours": {
                    "description": "How often to run cleanup",
                    "type": "integer"
                },
                "enabled": {
                    "description": "Enable automatic dead proxy cleanup",
                    "type": "boolean"
                },
                "max_failed_days": {
                    "description": "Remove proxies failed for more than N days",
                    "type": "integer"
                },
                "min_success_rate": {
                    "description": "Remove proxies with success rate below X% (0 = disabled)",
                    "type": "number"
                }
            }
        },
...
        "models.RateLimitSettings": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "interval": {
                    "description": "in seconds",
                    "type": "integer"
                },
                "max_requests": {
                    "type": "integer"
                }
            }
        },
...
        "models.RotationSettings": {
            "type": "object",
            "properties": {
                "allowed_protocols": {
                    "description": "[\"http\", \"https\", \"socks4\", \"socks4a\", \"socks5\"], empty means all",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fallback": {
                    "type": "boolean"
                },
                "fallback_max_retries": {
                    "type": "integer"
                },
                "follow_redirect": {
                    "type": "boolean"
                },
                "max_response_time": {
                    "description": "in milliseconds, 0 means no limit",
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "min_success_rate": {
                    "description": "0-100, 0 means no minimum",
                    "type": "number"
                },
                "remove_unhealthy": {
                    "type": "boolean"
                },
                "retries": {
                    "type": "integer"
                },
                "time_based": {
                    "$ref": "#/definitions/models.TimeBasedSettings"
                },
                "timeout": {
                    "type": "integer"
                }
            }
        },
        "models.Settings": {
            "type": "object",
            "properties": {
                "authentication": {
                    "$ref": "#/definitions/models.AuthenticationSettings"
                },
                "geoip": {
                    "$ref": "#/definitions/models.GeoIPSettings"
                },
                "global_health_check": {
                    "$ref": "#/definitions/models.GlobalHealthCheckSettings"
                },
                "healthcheck": {
                    "$ref": "#/definitions/models.HealthCheckSettings"
                },
                "log_retention": {
                    "$ref": "#/definitions/models.LogRetentionSettings"
                },
                "proxy_cleanup": {
                    "$ref": "#/definitions/models.ProxyCleanupSettings"
                },
                "rate_limit": {
                    "$ref": "#/definitions/models.RateLimitSettings"
                },
                "rotation": {
                    "$ref": "#/definitions/models.RotationSettings"
                }
            }
        },
...
        "models.TimeBasedSettings": {
            "type": "object",
            "properties": {
                "interval": {
                    "description": "in seconds",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}